add_metadata adds the remaining logic for the master branch to be able to run CR rejection on inputs from all six detectors. The pipeline now accepts _flt.fits files, in addition to _c0m.fits files.

CR rejection is run on WFPC2, HRC, WFC, and UVIS using parameters acquired from files unique to each detector in MTPipeline/cosmicx_cfg. For HRC, WFC, and UVIS, the readnoise and gain are read from the image header files and supplant the readnoise and gain set in the configuration files. CR rejection is not run on IR or SBC images. We are confident that no CR rejection is needed for our IR data, and the same likely holds true for SBC (though not for the same reason). At the moment, the cosmicx parameter files for IR and SBC remain, but can in the future be removed (although this will require changes into get_cosmicx_params).

+++ mtpipeline/imaging/run_cosmicx.py
@@ -4,6 +4,7 @@
 """
 
 import os
+import shutil
 import inspect
 import glob
 import yaml
@@ -35,17 +36,14 @@ def get_file_list(search_string):
 
 # -----------------------------------------------------------------------------
 
-def get_cosmicx_params(detector):
+def get_cosmicx_params(header_data):
     """ Accesses the cosmicx configure file and returns the parameters.
     
     Parameters:
         instrument: string
             The intrument for the image. Can be 'WFPC2', 'WFC3', 'ACS'
-        detector: string
-            The detector aboard the instrument. Can be:
-                None for WFPC2, 
-                'UVIS', 'IR' for WFC3
-                'SBC', 'HRC', 'WFC' for ACS
+        header_data: dictionary
+                Dictionary containing information from FITS header
 
     Returns:
         cosmicx_params: dict
@@ -57,7 +55,6 @@ def get_cosmicx_params(detector):
     """
 
 
-
     detector_config = { 'WFPC2' : "WFPC2_params.yaml",
                         'HRC' : "HRC_params.yaml",
                         'SBC' : "SBC_params.yaml",
@@ -65,36 +62,55 @@ def get_cosmicx_params(detector):
                         'UVIS' : "UVIS_params.yaml",
                         'IR' : "IR_params.yaml" }
 
+    detector = header_data["detector"]
+    readnoise = header_data["readnoise"]
+    gain = header_data["gain"]
+
+    # Load the configuration file for the proper detector:
+    config_file = detector_config[detector]
 
     # os.path.dirname is called three times to move up three directories
     # from this file, to reach to where the cosmicx_cfg directory is located.
     param_path = os.path.dirname(os.path.dirname(os.path.dirname(
                  os.path.abspath(inspect.getfile(inspect.currentframe())))))
-
     param_path = os.path.join(param_path,'cosmicx_cfg')
-    param_path = os.path.join(param_path,detector_config[detector])
+    param_path = os.path.join(param_path, config_file)
 
     cosmicx_params = yaml.load(open(param_path))
+
+    # If we have readnoise and gain from the FITS header, override the 
+    # manually specified values from the cfg file:
+    if readnoise:
+        for extension_key in cosmicx_params:
+            cosmicx_params[extension_key]["readnoise"] = readnoise
+    if gain:
+        for extension_key in cosmicx_params:
+            cosmicx_params[extension_key]["gain"] = gain
+
     return cosmicx_params 
 
 # -----------------------------------------------------------------------------
 
 
-def run_cosmicx(filename, output, cosmicx_params):
-    """ Driver to run lacosmicx on multi-extension WFPC2 FITS files.
+def run_cosmicx(filename, output, cosmicx_params, detector):
+    """ Driver to run lacosmicx on multi-extension FITS files.
 
     An equivalent to the run_cosmics function in run_cosmiscs.py,
     borrowing much of its code. Performs cosmic-ray rejection on each
-    chip of a FITS file from WFPC2.
+    chip of a FITS file.
 
     Parameters:
         filename: string
             filename of input FITS file
         output: string
             desired filename of output, a cosmic ray rejected FITS file.
-        iters: int
-            number of iterations of the LACosmics algorithm to be
-            performed.
+        cosmicx_params: dictionary
+            a dictionary of dictionaries, with each entry dictionary
+            containing the cosmicx settings appropriate for a
+            particular FITS science extension.
+        detector: string
+            the detector used to take the FITS image. If SBC or IR,
+            cr rejection is not run.
 
     Returns: nothing
 
@@ -102,16 +118,22 @@ def run_cosmicx(filename, output, cosmicx_params):
         A cosmic ray rejected FITS file with the filename given by
         'output'
         A symbolic link to the original c1m files that matches the
-        output filename, for astrodrizzle.
+        output filename, for Astrodrizzle.
 
     """
 
     # Assert the input file exists
-    error = filename + ' input for run_cosmics in '
+    error = filename + ' input for run_cosmicx in '
     error += 'run_cosmics.py does not exist.'
     assert os.access(filename, os.F_OK), error
 
-    # Define the output name and delete if exists.
+    # If the file is taken by SBC or IR, do not cr reject, copy
+    # the file with the output name, and return
+    if detector == 'SBC' or detector == 'IR':
+        shutil.copyfile(filename,output)
+        return
+
+    # Define the output name and delete if exists
     query = os.access(output, os.F_OK)
     if query == True:
         os.remove(output)
@@ -121,15 +143,26 @@ def run_cosmicx(filename, output, cosmicx_params):
         for key in cosmicx_params:
             
             params = cosmicx_params[key]
-            HDU = HDUlist[key]
+
+            # It's possible there may not be the same number of SCI extensions
+            # in the image as we specify in the cfg files (WFC)
+            try: 
+                HDU = HDUlist[key]
+            except IndexError:
+                continue
 
             # It's possible some of the SCI extensions might not have data,
             # as subsets of the CCD are sometimes used, and these extensions
-            # are empty.
+            # are empty (UVIS)
             try: 
                 array = HDU.data
             except AttributeError: 
                 continue
+
+            # For a cleaner cr rejection, set all very negative pixels
+            # to 0. No significant science data should be lost, as these
+            # are already bad pixels
+            array[array < -10.0] = 0.0
 
             cleanarray = lacosmicx.run(array,
                        inmask=params['inmask'],
@@ -156,7 +189,7 @@ def run_cosmicx(filename, output, cosmicx_params):
 
 def make_c1m_link(filename):
     """ Create a link to a c1m.fits that matches the cosmic ray rejected
-    naming scheme.
+    naming scheme. This is only done for WFPC2 data (which ends in _c0m.fits)
 
     Parameters:
         filename: string
@@ -168,11 +201,11 @@ def make_c1m_link(filename):
         A symbolic link to 'filename'
     """
 
-    error = filename + ' does not end in "c0m.fits".'
-    assert filename[-8:] == 'c0m.fits', error
-    src = filename.replace('_c0m.fits', '_c1m.fits')
-    dst = src.replace('_c1m.fits', '_cr_c1m.fits')
-    query = os.path.islink(dst)
-    if query == True:
-        os.remove(dst)
-    os.symlink(src, dst)
+    if filename[-8:] == 'c0m.fits':
+        src = filename.replace('_c0m.fits', '_c1m.fits')
+        dst = src.replace('_c1m.fits', '_cr_c1m.fits')
+        query = os.path.islink(dst)
+
+        if query == True:
+            os.remove(dst)
+        os.symlink(src, dst)
